More limit switch fixes
Fixed the logic for when we start out with a limit switch triggered in
both limit switch VIs.  Also fixed the issue of setting output in the
'normal' state, although I feel as though there is a better way to take
care of that, so that should get looked over.